+++ changelog.md
@@ -5,6 +5,9 @@
 - Change embed author constructor parameter order to `name, icon_url, url` from `icon_url`, `name`, `url`. On this
     way it is more intuitive and matches the footer's parameter order as well. Drop warning if passed in bad order.
 
+#### Bug fixes
+- `Sticker.user` was not updated when expected.
+
 ## 1.2.2 *\[2022-02-28\]*
 
 #### Improvements